PST-20499 prestashop   matching carrier

+++ classes/models/LengowInstall.php
@@ -190,8 +190,6 @@ class LengowInstall
         );
         $oldVersion = LengowConfiguration::getGlobalValue(LengowConfiguration::PLUGIN_VERSION);
         $oldVersion = $oldVersion ?: false;
-        // remove Lengow config
-        Db::getInstance()->execute('DELETE FROM `' . _DB_PREFIX_ . 'configuration` WHERE name LIKE "LENGOW_%"');
         $this->setDefaultValues();
         $this->update($oldVersion);
         LengowMain::log(
@@ -248,7 +246,9 @@ class LengowInstall
         // check if update is in progress
         self::setInstallationStatus(true);
         // remove Lengow config
-        Db::getInstance()->execute('DELETE FROM `' . _DB_PREFIX_ . 'configuration` WHERE name LIKE "LENGOW_%"');
+        if ($oldVersion) {
+            Db::getInstance()->execute('DELETE FROM `' . _DB_PREFIX_ . 'configuration` WHERE name LIKE "LENGOW_%"');
+        }
         // create all Lengow tables
         $this->createLengowTables();
         // run sql script and configuration upgrade for specific version
@@ -374,7 +374,7 @@ class LengowInstall
                 LengowLog::CODE_UNINSTALL,
                 LengowMain::setLogMessage('log.uninstall.table_dropped', ['name' => $table])
             );
-            Db::getInstance()->Execute('DROP TABLE IF EXISTS ' . _DB_PREFIX_ . $table);
+            Db::getInstance()->execute('DROP TABLE IF EXISTS ' . _DB_PREFIX_ . $table);
         }
         return true;
     }
